requirements.txt
azure-cognitiveservices-speech
boto3
ffmpeg
pydub
speechrecognition
